# chess_backend/interfaces/openapi.runtime.json
{"openapi":"3.1.0","info":{"title":"Online Chess Matchmaker API","description":"Backend for online chess play: matchmaking, move validation, persistence, and history.","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Health"],"summary":"Health check","description":"PUBLIC_INTERFACE\nBasic health check.\n\nReturns:\n    JSON message indicating the service is running.","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/healthz":{"get":{"tags":["Health"],"summary":"Detailed health check","description":"PUBLIC_INTERFACE\nHealth check that also validates database connectivity.\n\nReturns:\n    JSON with db_ok boolean and timestamp.","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/players":{"post":{"tags":["Players"],"summary":"Create or fetch player by nickname","description":"PUBLIC_INTERFACE\nCreate a new player with a unique nickname.\n\nIf the nickname already exists, returns the existing player.","operationId":"create_player_players_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayerCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matchmaking/enqueue":{"post":{"tags":["Matchmaking"],"summary":"Join matchmaking queue","description":"PUBLIC_INTERFACE\nEnqueue a player for matchmaking.\n\nReturns a ticket id. Client should poll /matchmaking/tickets/{ticket_id}.","operationId":"enqueue_matchmaking_matchmaking_enqueue_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchmakingEnqueueRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchmakingTicketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matchmaking/tickets/{ticket_id}":{"get":{"tags":["Matchmaking"],"summary":"Get matchmaking ticket status","description":"PUBLIC_INTERFACE\nPoll a matchmaking ticket.","operationId":"get_ticket_matchmaking_tickets__ticket_id__get","parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"string","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchmakingTicketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}":{"get":{"tags":["Games"],"summary":"Get game state","description":"PUBLIC_INTERFACE\nFetch current game state (FEN, status, players).","operationId":"get_game_games__game_id__get","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/with-moves":{"get":{"tags":["Games"],"summary":"Get game state including move list","description":"PUBLIC_INTERFACE\nFetch game state and full move list (chronological).","operationId":"get_game_with_moves_games__game_id__with_moves_get","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameWithMovesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/moves":{"post":{"tags":["Games"],"summary":"Submit a move (validated server-side)","description":"PUBLIC_INTERFACE\nSubmit a move for a game.\n\nValidations:\n- game exists and is active\n- player is in the game\n- it's player's turn\n- move is legal per current FEN\n\nPersists move and updates game current_fen and status if game ends.","operationId":"submit_move_games__game_id__moves_post","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/players/{player_id}/history":{"get":{"tags":["Games"],"summary":"Get recent games for a player","description":"PUBLIC_INTERFACE\nReturn recent games involving the specified player.","operationId":"player_history_players__player_id__history_get","parameters":[{"name":"player_id","in":"path","required":true,"schema":{"type":"string","title":"Player Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Max number of games to return.","default":20,"title":"Limit"},"description":"Max number of games to return."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HistoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GameResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Game UUID."},"white_player_id":{"type":"string","title":"White Player Id","description":"White player UUID."},"black_player_id":{"type":"string","title":"Black Player Id","description":"Black player UUID."},"status":{"type":"string","enum":["active","finished"],"title":"Status","description":"Game status."},"winner_color":{"anyOf":[{"type":"string","enum":["white","black"]},{"type":"null"}],"title":"Winner Color","description":"Winner if finished; null for draw/active."},"current_fen":{"type":"string","title":"Current Fen","description":"Current board state as FEN."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Last update timestamp."},"ended_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ended At","description":"End timestamp (if finished)."}},"type":"object","required":["id","white_player_id","black_player_id","status","current_fen","created_at","updated_at"],"title":"GameResponse"},"GameWithMovesResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Game UUID."},"white_player_id":{"type":"string","title":"White Player Id","description":"White player UUID."},"black_player_id":{"type":"string","title":"Black Player Id","description":"Black player UUID."},"status":{"type":"string","enum":["active","finished"],"title":"Status","description":"Game status."},"winner_color":{"anyOf":[{"type":"string","enum":["white","black"]},{"type":"null"}],"title":"Winner Color","description":"Winner if finished; null for draw/active."},"current_fen":{"type":"string","title":"Current Fen","description":"Current board state as FEN."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Last update timestamp."},"ended_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Ended At","description":"End timestamp (if finished)."},"moves":{"items":{"$ref":"#/components/schemas/MoveResponse"},"type":"array","title":"Moves","description":"Moves in chronological order."}},"type":"object","required":["id","white_player_id","black_player_id","status","current_fen","created_at","updated_at","moves"],"title":"GameWithMovesResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryResponse":{"properties":{"games":{"items":{"$ref":"#/components/schemas/GameResponse"},"type":"array","title":"Games","description":"Recent games for player."}},"type":"object","required":["games"],"title":"HistoryResponse"},"MatchmakingEnqueueRequest":{"properties":{"player_id":{"type":"string","title":"Player Id","description":"Player UUID."}},"type":"object","required":["player_id"],"title":"MatchmakingEnqueueRequest"},"MatchmakingTicketResponse":{"properties":{"ticket_id":{"type":"string","title":"Ticket Id","description":"Matchmaking ticket UUID."},"status":{"type":"string","enum":["queued","matched","cancelled"],"title":"Status","description":"Current ticket status."},"game_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Game Id","description":"Game UUID when matched."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Ticket creation time."},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Last update time."}},"type":"object","required":["ticket_id","status","created_at","updated_at"],"title":"MatchmakingTicketResponse"},"MoveCreateRequest":{"properties":{"player_id":{"type":"string","title":"Player Id","description":"Player UUID making the move."},"uci":{"type":"string","maxLength":5,"minLength":4,"title":"Uci","description":"Move in UCI, e.g. e2e4, e7e8q."}},"type":"object","required":["player_id","uci"],"title":"MoveCreateRequest"},"MoveResponse":{"properties":{"id":{"type":"integer","title":"Id","description":"Move row id."},"game_id":{"type":"string","title":"Game Id","description":"Game UUID."},"move_number":{"type":"integer","title":"Move Number","description":"Move number (1-based)."},"color":{"type":"string","enum":["white","black"],"title":"Color","description":"Color that moved."},"uci":{"type":"string","title":"Uci","description":"UCI string."},"san":{"type":"string","title":"San","description":"SAN notation."},"fen_after":{"type":"string","title":"Fen After","description":"FEN after move."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."}},"type":"object","required":["id","game_id","move_number","color","uci","san","fen_after","created_at"],"title":"MoveResponse"},"PlayerCreateRequest":{"properties":{"nickname":{"type":"string","maxLength":24,"minLength":2,"title":"Nickname","description":"Public nickname (unique)."}},"type":"object","required":["nickname"],"title":"PlayerCreateRequest"},"PlayerResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Player UUID."},"nickname":{"type":"string","title":"Nickname","description":"Public nickname."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."}},"type":"object","required":["id","nickname","created_at"],"title":"PlayerResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Health","description":"Service health and basic diagnostics."},{"name":"Players","description":"Create and manage player identities."},{"name":"Matchmaking","description":"Queue players and match them into games."},{"name":"Games","description":"Game state, moves, and history."}]}